+++ AdapterUnitTests/AdapterUnitTestKit/AUTKConstants.h
@@ -1,4 +1,0 @@
-#import <Foundation/Foundation.h>
-
-/// Default timeout for XC expectation.
-extern NSTimeInterval const AUTKExpectationTimeout;