+++ adapters/Chartboost/ChartboostAdapter/GADMAdapterChartboostConstants.h
@@ -15,7 +15,7 @@
 #import <Foundation/Foundation.h>
 
 /// Chartboost mediation network adapter version.
-static NSString *const GADMAdapterChartboostVersion = @"9.3.0.0";
+static NSString *const GADMAdapterChartboostVersion = @"9.4.0.0";
 
 /// Chartboost App ID.
 static NSString *const GADMAdapterChartboostAppID = @"appId";