+++ adapters/InMobi/InMobiAdapter/GADMAdapterInMobiConstants.h
@@ -12,8 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#import <Foundation/Foundation.h>
+
 /// Adapter version string.
-static NSString *const GADMAdapterInMobiVersion = @"10.5.5.0";
+static NSString *const GADMAdapterInMobiVersion = @"10.5.6.0";
 
 /// InMobi SDK key.
 static NSString *const GADMAdapterInMobiAccountID = @"accountid";
@@ -23,3 +25,26 @@ static NSString *const GADMAdapterInMobiPlacementID = @"placementid";
 
 /// InMobi adapter error domain.
 static NSString *const GADMAdapterInMobiErrorDomain = @"com.google.mediation.inmobi";
+
+/// IAB U.S. Privacy string key.
+static NSString *const GADMAdapterInMobiIABUSPrivacyString = @"IABUSPrivacy_String";
+
+/// Key for the InMobi request parameters SDK version.
+static NSString *const GADMAdapterInMobiRequestParametersSDKVersionKey = @"tp-ver";
+
+/// Key for the InMobi request parameters mediation type.
+static NSString *const GADMAdapterInMobiRequestParametersMediationTypeKey = @"tp";
+
+/// Key for the InMobi request parameters COPPA.
+static NSString *const GADMAdapterInMobiRequestParametersCOPPAKey = @"coppa";
+
+/// Type of InMobi mediation.
+typedef NSString *GADMAdapterInMobiRequestParametersMediationType NS_TYPED_ENUM;
+
+/// InMobi waterfall mediation type.
+static GADMAdapterInMobiRequestParametersMediationType const
+    GADMAdapterInMobiRequestParametersMediationTypeWaterfall = @"c_admob";
+
+/// InMobi RTB mediation type.
+static GADMAdapterInMobiRequestParametersMediationType const
+    GADMAdapterInMobiRequestParametersMediationTypeRTB = @"c_google";