+++ adapters/IronSource/IronSourceAdapter/GADMAdapterIronSourceConstants.h
@@ -14,7 +14,9 @@
 
 // IronSource internal reporting const.
 static NSString *const GADMAdapterIronSourceMediationName = @"AdMob";
-static NSString *const GADMAdapterIronSourceAdapterVersion = @"7.3.1.0";
+
+static NSString *const GADMAdapterIronSourceAdapterVersion = @"7.4.0.0.0";
+
 static NSString *const GADMIronSourceDefaultInstanceId = @"0";
 static NSString *const GADMAdapterIronSourceInternalVersion = @"400";