+++ adapters/Line/LineAdapter/GADMediationAdapterLineConstants.m
@@ -12,4 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#import "GADMediationAdapterLineExtras.h"
+#import "GADMediationAdapterLineConstants.h"
+
+NSString *const GADMediationAdapterLineVersion = @"2.6.20230609.0";
+
+NSString *const GADMediationAdapterLineCredentialKeyApplicationID = @"application_id";
+
+NSString *const GADMediationAdapterLineCredentialKeyAdUnit = @"slot_id";
+
+NSString *const GADMediationAdapterLineErrorDomain = @"com.google.mediation.line";
+
+NSString *const GADMediationAdapterFiveAdErrorDomain = @"com.google.mediation.fiveadsdk";