+++ adapters/Line/LineAdapter/GADMediationAdapterLineExtras.m
@@ -13,3 +13,6 @@
 // limitations under the License.
 
 #import "GADMediationAdapterLineExtras.h"
+
+@implementation GADMediationAdapterLineExtras
+@end