+++ adapters/Pangle/PangleAdapter/GADMediationAdapterPangleConstants.h
@@ -15,7 +15,7 @@
 #import <Foundation/Foundation.h>
 
 /// Pangle adapter version.
-static NSString *const GADMAdapterPangleVersion = @"5.2.0.9.0";
+static NSString *const GADMAdapterPangleVersion = @"5.3.1.2.0";
 
 /// Pangle adapter Pangle App ID parameter key.
 static NSString *const GADMAdapterPangleAppID = @"appid";