+++ adapters/AppLovin/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## AppLovin iOS Mediation Adapter Changelog
 
+#### [Version 11.11.2.0](https://dl.google.com/googleadmobadssdk/mediation/ios/applovin/AppLovinAdapter-11.11.2.0.zip)
+- Verified compatibility with AppLovin SDK 11.11.2.
+
+Built and tested with:
+- Google Mobile Ads SDK version 10.9.0.
+- AppLovin SDK version 11.11.2.
+
 #### [Version 11.10.1.0](https://dl.google.com/googleadmobadssdk/mediation/ios/applovin/AppLovinAdapter-11.10.1.0.zip)
 - Verified compatibility with AppLovin SDK 11.10.1.
 - Removed support of the `armv7` architecture.