+++ adapters/DTExchange/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## DT Exchange iOS Mediation Adapter Changelog
 
+#### [Version 8.2.3.0](https://dl.google.com/googleadmobadssdk/mediation/ios/dtexchange/DTExchangeAdapter-8.2.3.0.zip)
+- Verified compatibility with DT Exchange SDK version 8.2.3.
+
+Built and tested with:
+- Google Mobile Ads SDK version 10.7.0.
+- DT Exchange SDK version 8.2.3.
+
 #### [Version 8.2.2.0](https://dl.google.com/googleadmobadssdk/mediation/ios/dtexchange/DTExchangeAdapter-8.2.2.0.zip)
 - Verified compatibility with DT Exchange SDK version 8.2.2.