+++ adapters/InMobi/CHANGELOG.md
@@ -1,6 +1,14 @@
 ## InMobi iOS Mediation Adapter Changelog
 
-#### Version 10.5.5.0 (In progress)
+#### [Version 10.5.6.0](https://dl.google.com/googleadmobadssdk/mediation/ios/inmobi/InMobiAdapter-10.5.6.0.zip)
+- Added support to read IAB U.S. Privacy string from `NSUserDefaults`.
+- Verified compatibility with InMobi SDK 10.5.6.
+
+Built and tested with:
+- Google Mobile Ads SDK version 10.9.0.
+- InMobi SDK version 10.5.6.
+
+#### [Version 10.5.5.0](https://dl.google.com/googleadmobadssdk/mediation/ios/inmobi/InMobiAdapter-10.5.5.0.zip)
 - Now requires Google Mobile Ads SDK version 10.4.0 or higher.
 - Updated the adapter to use the `didRewardUser` API.
 - Updated the adapter to initialize InMobi SDK on main thread.