+++ adapters/LiftoffMonetize/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Liftoff Monetize iOS Mediation Adapter Changelog
 
+#### [Version 7.0.1.0](https://dl.google.com/googleadmobadssdk/mediation/ios/liftoffmonetize/LiftoffMonetizeAdapter-7.0.1.0.zip)
+- Added watermark support for bidding ads.
+- Verified compatibility with Vungle SDK 7.0.1.
+
+Built and tested with:
+- Google Mobile Ads SDK version 10.7.0.
+- Vungle SDK version 7.0.1.
+
 #### [Version 7.0.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/liftoffmonetize/LiftoffMonetizeAdapter-7.0.0.0.zip)
 - Rebranded adapter name from Vungle to Liftoff Monetize.
 - Verified compatibility with Vungle SDK 7.0.0.