+++ adapters/Unity/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Unity Ads iOS Mediation Adapter Changelog
 
+#### [Version 4.8.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/unity/UnityAdapter-4.8.0.0.zip)
+- Added support for impression events for banner ads.
+- Verified compatibility with Unity Ads SDK 4.8.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 10.7.0.
+- Unity Ads SDK version 4.8.0.
+
 #### [Version 4.7.1.0](https://dl.google.com/googleadmobadssdk/mediation/ios/unity/UnityAdapter-4.7.1.0.zip)
 - Verified compatibility with Unity Ads SDK 4.7.1.